refactor: split decl stmt handlers
